==> jupyter_notebooks/01 - DataCollection.ipynb <==
# **Data Collection**

## Objectives

* Fetch data from Kaggle and save it as raw data.
* Inspect the data and save it under outputs/datasets/collection

## Inputs

* Kaggle JSON file - the authentication token.
* Kaggle dataset URL - fedesoriano/heart-failure-prediction/data

## Outputs

* outputs/datasets/collection/HeartDiseasePrediction.csv

## Additional Comments

* In case you have any additional comments that don't fit in the previous bullets, please state them here. 


---

# Change working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os


current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Import Dataset from Kaggle

Firstly, the Kaggle API must be installed before the data can be loaded.

A valid account must be registered with Kaggle to obtain an API key (as a JSON-file).

! pip install kaggle==1.6.14

Next, the Kaggle config directory is set to the current working directory, and the read/write permissions are set to user only (600)

os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

Then, define the Kaggle dataset and destination folder paths and download to 'inputs/datasets/raw' directory.

* The dataset path is taken from the Kaggle url, after 'https://www.kaggle.com/datasets/'

KaggleDatasetPath = "fedesoriano/heart-failure-prediction"
DestinationFolder = "inputs/datasets/raw"
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

The downloaded file is then unzipped, and the zipped file and kaggle.json are both deleted.

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
    && rm {DestinationFolder}/*.zip \
        && rm kaggle.json

---

# Load and Inspect the Kaggle Data

Using the pandas library, the dataset can be loaded as a dataframe and the data inspected.

import pandas as pd


df = pd.read_csv(f"inputs/datasets/raw/heart.csv")
df.head()

A summary of the dataframe columns, non-null counts and datatypes can be obtained.

There are 918 observations in the dataset, so the number of missing data can be calculated by 918 - non-null count.

df.info()

Each column has 918 non-null count, so it looks like there is no missing data. It is, however, good practice to double check.

If there are no missing data, we expect an empty list.

columns_with_nan = df.columns[df.isna().sum() > 0].to_list()
columns_with_nan

There are no ID columns, so there is no need to check for any duplicate values.

The FastingBS feature is currently a numerical datatype, however we know it should be categorical so this needs to be converted.

* As a reminder of the FastingBs feature: 1 = fasting blood sugar > 120 mg/dL, 0 = otherwise.
* A fasting blood sugar value of >125 mg/dL is typically indicative of diabetes, therefore we will map 1 to 'high diabetes risk' and 0 to 'low diabetes risk'

fastingbs_map = {0: "low diabetes risk", 1: "high diabetes risk"}
df["FastingBS"].replace(to_replace=fastingbs_map, inplace=True)
df["FastingBS"]

Confirming that the FastingBS feature's datatype has changed.

df.dtypes

---

# Save the Dataset

The modified dataset is saved to the outputs directory.

try:
  os.makedirs(name="outputs/datasets/collection")
except Exception as e:
  print(e)

df.to_csv(f"outputs/datasets/collection/HeartDiseasePrediction.csv",index=False)

---

# Conclusions

In this notebook, the following was achieved:

* The dataset was successfully imported via the Kaggle API
* The dataset was inspected, and no missing values were found
* The "FastingBS" feature was mapped to categorical values and it's datatype was confirmed to change
* The dataset was saved in the outputs directory

## Next Steps

In the next notebook, a exploratory data analysis will be carried out using Pandas profiling and correlation studies.

This work will make up the **'Data Understanding'** aspect of the CRISP-DM workflow, and will provide further understanding of the dataset and address business requirement 1.
